moving azure to _azure to avoid namespace collisions
