+++ python/ray/autoscaler/_private/_azure/azure-config-template.json
