+++ python/ray/autoscaler/_private/_azure/azure-vm-template.json
