+++ python/ray/autoscaler/_private/_azure/node_provider.py
