+++ python/ray/tests/test_autoscaler_yaml.py
@@ -11,8 +11,8 @@ from unittest.mock import MagicMock, Mock, patch
 import pytest
 from click.exceptions import ClickException
 
-from ray.autoscaler._private.azure.config import (_configure_key_pair as
-                                                  _azure_configure_key_pair)
+from ray.autoscaler._private._azure.config import (_configure_key_pair as
+                                                   _azure_configure_key_pair)
 from ray.autoscaler._private.gcp import config as gcp_config
 from ray.autoscaler._private.util import prepare_config, validate_config,\
     _get_default_config, merge_setup_commands
